% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnuplot.R
\name{scale_color_gnuplot}
\alias{scale_color_gnuplot}
\alias{gnucolors}
\alias{gnupalette}
\title{gnuplot's default color palette for discrete data, ported to ggplot2}
\usage{
scale_color_gnuplot(..., na.value = "grey50", aesthetics = "color")
gnupalette(n)
}
\arguments{
\item{...}{Passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{na.value}{Colour to use for missing values}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{n}{The number of colors to return from the palette}
}
\description{
These functions provide gnuplot's default color palette. Use the function
\code{scale_color_gnuplot()} with ggplot2, and the function \code{gnupalette()} or the
vector \code{gnucolors} otherwise.
}
\details{
Attention: Do not use this palette with more than eight categories, as the
colors are recycled on the ninth category!
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point() +
  scale_color_gnuplot() +
  scale_x_gnuplot() +
  scale_y_gnuplot() +
  theme_gnuplot()

gnupalette(3)
gnucolors[1:3]
}
\seealso{
\code{\link[ggplot2:scale_hue]{ggplot2::scale_color_discrete()}}
}
