% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnuplot.R
\name{scale_x_gnuplot}
\alias{scale_x_gnuplot}
\alias{gnubreaks}
\alias{gnulabels}
\alias{scale_y_gnuplot}
\title{gnuplot axes for ggplot2}
\usage{
scale_x_gnuplot(nbreaks = 5, padding = 0.1, breaks, labels, sec.axis, ...)
scale_y_gnuplot(nbreaks = 5, padding = 0.1, breaks, labels, sec.axis, ...)
gnubreaks(nbreaks = 5, padding = 0.1)
gnulabels()
}
\arguments{
\item{nbreaks}{The number of breaks/ticks on the axis}

\item{padding}{The amount of space between the outmost ticks and the
plot borders relative to the plot width. A number between
0 and 0.5.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
transformation object
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output
}}

\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}

\item{sec.axis}{specify a secondary axis}

\item{...}{Passed on to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_*_continuous()}}}
}
\description{
These functions try to choose pretty axis breaks/ticks and labels.
They also set up secondary axes.
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point() +
  scale_color_gnuplot() +
  scale_x_gnuplot() +
  scale_y_gnuplot() +
  theme_gnuplot()

# Alternatively, without the secondary axis:
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point() +
  scale_color_gnuplot() +
  scale_x_gnuplot(sec.axis = waiver()) +
  scale_y_gnuplot(sec.axis = waiver()) +
  theme_gnuplot()

breaks <- gnubreaks()(limits = c(0, 1000))
gnulabels()(breaks)
}
