% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnuplot.R
\name{scale_color_gnuplot}
\alias{scale_color_gnuplot}
\alias{gnucolors}
\alias{gnupalette}
\title{gnuplot's default color palette for discrete data, ported to ggplot2}
\usage{
scale_color_gnuplot(..., na.value = "grey50", aesthetics = "color")
gnupalette(n)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale}}
  \describe{
    \item{\code{scale_name}}{The name of the scale}
    \item{\code{palette}}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take.}
    \item{\code{name}}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
transformation object
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
    \item{\code{labels}}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
    \item{\code{limits}}{A character vector that defines possible values of the scale
and their order.}
    \item{\code{expand}}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
    \item{\code{na.translate}}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
    \item{\code{drop}}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} uses all the levels in the factor.}
    \item{\code{guide}}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more info.}
    \item{\code{position}}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}

\item{na.value}{Colour to use for missing values}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{n}{The number of colors to return from the palette}
}
\description{
These functions provide gnuplot's default color palette. Use the function
\code{scale_color_gnuplot()} with ggplot2, and the function \code{gnupalette()} or the
vector \code{gnucolors} otherwise.
}
\details{
Attention: Do not use this palette with more than eight categories, as the
colors are recycled on the ninth category!
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point() +
  scale_color_gnuplot() +
  scale_x_gnuplot() +
  scale_y_gnuplot() +
  theme_gnuplot()

gnupalette(3)
gnucolors[1:3]
}
\seealso{
\code{\link[ggplot2:scale_color_hue]{ggplot2::scale_color_hue()}}
}
